pyproject.toml
[project]
name = "github-blog"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "geoqiao", email = "realvincentqiao@gmail.com" }
]
dependencies = [
    "Jinja2",
    "PyGithub",
    "feedgen",
    "marko>=2.0.3",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
]
